+++ ForensicWhisperDeskZH/Document/IDocumentService.cs
@@ -13,18 +13,18 @@ namespace ForensicWhisperDeskZH.Document
         /// <param name="text">Text to insert</param>
         /// <returns>True if successful, false otherwise</returns>
         bool InsertText(string text);
-        
+
         /// <summary>
         /// Checks if a document is available for text insertion
         /// </summary>
         bool IsDocumentAvailable { get; }
-        
+
         /// <summary>
         /// Occurs when an error happens during document operations
         /// </summary>
         event EventHandler<DocumentErrorEventArgs> Error;
     }
-    
+
     /// <summary>
     /// Arguments for document error events
     /// </summary>
@@ -32,7 +32,7 @@ namespace ForensicWhisperDeskZH.Document
     {
         public string Message { get; }
         public Exception Exception { get; }
-        
+
         public DocumentErrorEventArgs(string message, Exception exception = null)
         {
             Message = message;
