+++ ForensicWhisperDeskZH/Text/LoggingService.cs
@@ -16,12 +16,12 @@ namespace ForensicWhisperDeskZH.Common
             Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData),
             "FennecTranscriptionSystem",
             "Logs");
-            
+
         private static readonly string LogPath = Path.Combine(LogDirectory, $"log_{DateTime.Now:yyyyMMdd}.txt");
         private static readonly string ErrorPath = Path.Combine(LogDirectory, $"errors_{DateTime.Now:yyyyMMdd}.txt");
-        
+
         private static readonly SemaphoreSlim LogLock = new SemaphoreSlim(1, 1);
-        
+
         static LoggingService()
         {
             try
@@ -33,7 +33,7 @@ namespace ForensicWhisperDeskZH.Common
                 // Fallback to local directory if unable to create log directory
             }
         }
-        
+
         /// <summary>
         /// Logs an informational message
         /// </summary>
@@ -41,7 +41,7 @@ namespace ForensicWhisperDeskZH.Common
         {
             await LogToFileAsync(LogPath, $"[INFO] [{DateTime.Now:yyyy-MM-dd HH:mm:ss}] [{source}] {message}");
         }
-        
+
         /// <summary>
         /// Logs an error message and optional exception
         /// </summary>
@@ -50,17 +50,17 @@ namespace ForensicWhisperDeskZH.Common
             var sb = new StringBuilder();
             sb.AppendLine($"[ERROR] [{DateTime.Now:yyyy-MM-dd HH:mm:ss}] [{source}]");
             sb.AppendLine($"Message: {message}");
-            
+
             if (ex != null)
             {
                 sb.AppendLine("Exception Details:");
                 AppendExceptionDetails(sb, ex);
             }
-            
+
             await LogToFileAsync(ErrorPath, sb.ToString());
             await LogToFileAsync(LogPath, $"[ERROR] [{DateTime.Now:yyyy-MM-dd HH:mm:ss}] [{source}] {message}");
         }
-        
+
         /// <summary>
         /// Logs a warning message
         /// </summary>
@@ -68,24 +68,24 @@ namespace ForensicWhisperDeskZH.Common
         {
             await LogToFileAsync(LogPath, $"[WARN] [{DateTime.Now:yyyy-MM-dd HH:mm:ss}] [{source}] {message}");
         }
-        
+
         private static void AppendExceptionDetails(StringBuilder sb, Exception ex, int level = 0)
         {
             if (ex == null) return;
-            
+
             string indent = new string(' ', level * 2);
             sb.AppendLine($"{indent}Type: {ex.GetType().FullName}");
             sb.AppendLine($"{indent}Message: {ex.Message}");
             sb.AppendLine($"{indent}StackTrace:");
             sb.AppendLine($"{indent}{ex.StackTrace}");
-            
+
             if (ex.InnerException != null)
             {
                 sb.AppendLine($"{indent}--- Inner Exception ---");
                 AppendExceptionDetails(sb, ex.InnerException, level + 1);
             }
         }
-        
+
         private static async Task LogToFileAsync(string path, string content)
         {
             try
@@ -108,18 +108,18 @@ namespace ForensicWhisperDeskZH.Common
                 // Unable to log, but we don't want to crash the application
             }
         }
-        
+
         // Synchronous versions for compatibility with existing code
         public static void LogMessage(string message, string source = "Application")
         {
             Task.Run(() => LogMessageAsync(message, source)).Wait();
         }
-        
+
         public static void LogError(string message, Exception ex = null, string source = "Application")
         {
             Task.Run(() => LogErrorAsync(message, ex, source)).Wait();
         }
-        
+
         public static void LogWarning(string message, string source = "Application")
         {
             Task.Run(() => LogWarningAsync(message, source)).Wait();
