+++ ForensicWhisperDeskZH/Transcription/ITranscriptionService.cs
@@ -14,27 +14,27 @@ namespace ForensicWhisperDeskZH.Transcription
         /// Occurs when transcription begins
         /// </summary>
         event EventHandler<TranscriptionEventArgs> TranscriptionStarted;
-        
+
         /// <summary>
         /// Occurs when transcription ends
         /// </summary>
         event EventHandler<TranscriptionEventArgs> TranscriptionStopped;
-        
+
         /// <summary>
         /// Occurs when a new piece of transcription text is available
         /// </summary>
         event EventHandler<TranscriptionResultEventArgs> TranscriptionResult;
-        
+
         /// <summary>
         /// Occurs when an error happens during transcription
         /// </summary>
         event EventHandler<ErrorEventArgs> TranscriptionError;
-        
+
         /// <summary>
         /// Gets whether the service is actively transcribing
         /// </summary>
         bool IsTranscribing { get; }
-        
+
         /// <summary>
         /// Starts transcription using the specified microphone
         /// </summary>
@@ -42,25 +42,29 @@ namespace ForensicWhisperDeskZH.Transcription
         /// <param name="deviceNumber">Microphone device number</param>
         /// <param name="language">Optional language to use</param>
         void StartTranscription(Action<string> textHandler, int deviceNumber = 0, CultureInfo language = null);
-        
+
         /// <summary>
         /// Stops active transcription
         /// </summary>
         void StopTranscription();
-        
+
         /// <summary>
         /// Toggles transcription on/off
         /// </summary>
         /// <param name="textHandler">Handler for transcribed text</param>
         /// <param name="deviceNumber">Microphone device number</param>
         void ToggleTranscription(Action<string> textHandler, TranscriptionSettings transcriptionSettings, int deviceNumber = 0);
-        
+
         /// <summary>
         /// Changes the language used for transcription
         /// </summary>
         /// <param name="language">Two-letter language code</param>
         void ChangeLanguage(string language);
 
+
+        /// <summary>
+        /// Changes the Model Type used for transcription
+        /// </summary>
         void ChangeModelType(GgmlType modelType);
 
     }
