+++ ForensicWhisperDeskZH/Transcription/TextProcessor.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Text.RegularExpressions;
 
 namespace ForensicWhisperDeskZH.Transcription
 {
@@ -9,12 +8,12 @@ namespace ForensicWhisperDeskZH.Transcription
     public class TextProcessor
     {
         private readonly TranscriptionSettings _settings;
-        
+
         public TextProcessor(TranscriptionSettings settings)
         {
             _settings = settings ?? throw new ArgumentNullException(nameof(settings));
         }
-        
+
         /// <summary>
         /// Formats transcribed text according to settings (capitalization, punctuation)
         /// </summary>
@@ -28,7 +27,7 @@ namespace ForensicWhisperDeskZH.Transcription
 
             return result;
         }
-        
+
         /// <summary>
         /// Extracts only the new text portion compared to previous text
         /// </summary>
@@ -44,7 +43,7 @@ namespace ForensicWhisperDeskZH.Transcription
 
             // Find overlap between the end of previous text and start of new text
             int overlap = FindMaxOverlap(previousText, newText);
-            
+
             // If we found an overlap, return the non-overlapping part
             if (overlap > 0)
                 return newText.Substring(overlap);
@@ -52,14 +51,14 @@ namespace ForensicWhisperDeskZH.Transcription
             // If no overlap, it's completely new text
             return newText;
         }
-        
+
         /// <summary>
         /// Finds the maximum overlap between the end of string1 and start of string2
         /// </summary>
         private int FindMaxOverlap(string string1, string string2)
         {
             int maxOverlap = 0;
-            
+
             for (int i = 1; i <= Math.Min(string1.Length, string2.Length); i++)
             {
                 string prevEnd = string1.Substring(string1.Length - i);
@@ -68,7 +67,7 @@ namespace ForensicWhisperDeskZH.Transcription
                 if (string.Equals(prevEnd, newStart, StringComparison.OrdinalIgnoreCase))
                     maxOverlap = i;
             }
-            
+
             return maxOverlap;
         }
     }
